requirements.txt
Flask==1.1.4
Flask_Assets==2.0
python-dotenv==0.19.2
